fix: zero ltv warning (#2073)
