+++ src/ui-config/errorMapping.tsx
@@ -130,7 +130,7 @@ export const errorMapping: Record<number, ReactElement> = {
   // 59: <Trans>Price oracle sentinel validation failed</Trans>,
   60: <Trans>Asset is not borrowable in isolation mode</Trans>,
   // 61: <Trans>Reserve has already been initialized</Trans>,
-  62: <Trans>User is in isolation mode</Trans>,
+  62: <Trans>User is in isolation mode or LTV is zero</Trans>,
   // 63: <Trans>Invalid ltv parameter for the reserve</Trans>,
   // 64: <Trans>Invalid liquidity threshold parameter for the reserve</Trans>,
   // 65: <Trans>Invalid liquidity bonus parameter for the reserve</Trans>,
